clippy: automated corrections

+++ src/archive.rs
@@ -70,7 +70,7 @@ impl Slab {
         w.write_u64::<LittleEndian>(self.blocks.len() as u64)?;
         for b in &self.blocks {
             w.write_all(&b.h[..])?;
-            w.write_u32::<LittleEndian>(b.offset as u32)?;
+            w.write_u32::<LittleEndian>(b.offset)?;
         }
 
         let compressed = self.packer.complete()?;
