+++ src/stream_builders.rs
@@ -279,7 +279,7 @@ impl DeltaBuilder {
                 nr_entries,
             } => (
                 Partial {
-                    begin: *begin as u32,
+                    begin: *begin,
                     end: *begin + split_point as u32,
                     slab: *slab,
                     offset: *offset,
